change ListItem from factory to class; remove list_nav from tags&topics
