+++ dispatch/static/manager/src/js/actions/TagsActions.js
@@ -1,5 +1,4 @@
 import { push } from 'react-router-redux'
-import { normalize, arrayOf } from 'normalizr'
 
 import * as types from '../constants/ActionTypes'
 import { tagSchema } from '../constants/Schemas'
@@ -9,16 +8,6 @@ import DispatchAPI from '../api/dispatch'
 import { ResourceActions } from '../util/redux'
 
 class TagsActions extends ResourceActions {
-
-  listNav(token) {
-    return {
-      type: this.types.LIST_NAV,
-      payload: this.api.list(token)
-        .then(json => ({
-          data: normalize(json.results, arrayOf(this.schema))
-        }))
-    }
-  }
 
   search(query) {
     let queryObj = {}
