--- dispatch/static/manager/src/js/pages/ListItemFactory/ListItemPage.js
@@ -1,8 +1,0 @@
-import React from 'react'
-
-export default function ListItemPage(props) {
-  return React.createElement(props.editorClass, {
-    itemId: props.params.itemId,
-    goBack: props.history.goBack
-  })
-}
