--- dispatch/static/manager/src/js/pages/ListItemFactory/NewListItemPage.js
@@ -1,8 +1,0 @@
-import React from 'react'
-
-export default function NewListItemPage(props) {
-  return React.createElement(props.editorClass, {
-    isNew: true,
-    goBack: props.history.goBack
-  })
-}
