--- dispatch/static/manager/src/js/pages/ListItemFactory/index.js
@@ -1,9 +1,0 @@
-import ListItemIndexPageFactory from './ListItemIndexPageFactory'
-import ListItemPage from './ListItemPage'
-import NewListItemPage from './NewListItemPage'
-
-export {
-  ListItemIndexPageFactory as IndexFactory,
-  ListItemPage as ListItem,
-  NewListItemPage as NewListItem
-}
